Update the stored procedure helpers in the TPCC benchmark
