+++ src/main/java/org/vanilladb/bench/server/param/tpcc/TpccSchemaBuilderSpParamHelper.java
@@ -1,0 +1,46 @@
+/*******************************************************************************
+ * Copyright 2016, 2017 vanilladb.org contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *******************************************************************************/
+package org.vanilladb.bench.server.param.tpcc;
+
+import org.vanilladb.bench.benchmarks.tpcc.rte.TpccSchemaBuilderParamHelper;
+import org.vanilladb.core.sql.Schema;
+import org.vanilladb.core.sql.storedprocedure.SpResultRecord;
+import org.vanilladb.core.sql.storedprocedure.StoredProcedureParamHelper;
+
+public class TpccSchemaBuilderSpParamHelper extends TpccSchemaBuilderParamHelper
+		implements StoredProcedureParamHelper {
+
+
+	@Override
+	public void prepareParameters(Object... pars) {
+		// nothing to do
+	}
+
+	@Override
+	public Schema getResultSetSchema() {
+		return new Schema();
+	}
+
+	@Override
+	public SpResultRecord newResultSetRecord() {
+		return new SpResultRecord();
+	}
+
+	@Override
+	public boolean isReadOnly() {
+		return false;
+	}
+}
